Make testable setup
