+++ Implementation/src/test/java/com/itestra/software_analyse_challenge/LineNumberCounterTest.java
@@ -1,0 +1,4 @@
+package com.itestra.software_analyse_challenge;
+
+public class LineNumberCounterTest {
+}
